+++ example/example.dart
@@ -22,6 +22,10 @@ void main() {
   print('----- RSA-PSS SHA-256 -----');
   ps256();
   print('---------------------------\n');
+
+  print('----- RSA Certificate -----');
+  rsaCert();
+  print('---------------------------\n');
 }
 
 // HMAC SHA-256 algorithm
@@ -221,6 +225,49 @@ void ps256() {
       // Verify a token
       final pem = File('./example/rsa_pss_public.pem').readAsStringSync();
       final key = RSAPublicKey(pem);
+
+      final jwt = JWT.verify(token, key);
+
+      print('Payload: ${jwt.payload}');
+    } on JWTExpiredException {
+      print('jwt expired');
+    } on JWTException catch (ex) {
+      print(ex.message); // ex: invalid signature
+    }
+  }
+}
+
+// RSA Certificate
+void rsaCert() {
+  String token;
+
+  /* Sign */ {
+    // Create a json web token
+    final jwt = JWT(
+      {
+        'id': 123,
+        'server': {
+          'id': '3e4fc296',
+          'loc': 'euw-2',
+        }
+      },
+      issuer: 'https://github.com/jonasroussel/dart_jsonwebtoken',
+    );
+
+    // Sign it
+    final pem = File('./example/rsa_cert_private.pem').readAsStringSync();
+    final key = RSAPrivateKey(pem);
+
+    token = jwt.sign(key, algorithm: JWTAlgorithm.RS256);
+
+    print('Signed token: $token\n');
+  }
+
+  /* Verify */ {
+    try {
+      // Verify a token
+      final pem = File('./example/rsa_certificate.pem').readAsStringSync();
+      final key = RSAPublicKey.cert(pem);
 
       final jwt = JWT.verify(token, key);